fix: 앱 설치 모달 출력
